+++ cucumber-expressions/java/src/main/java/io/cucumber/cucumberexpressions/RegularExpression.java
@@ -1,7 +1,5 @@
 package io.cucumber.cucumberexpressions;
 
-import org.apiguardian.api.API;
-
 import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
@@ -9,8 +7,7 @@ import java.util.regex.Pattern;
 
 import static io.cucumber.cucumberexpressions.ParameterType.createAnonymousParameterType;
 
-@API(status = API.Status.STABLE)
-public final class RegularExpression implements Expression {
+final class RegularExpression implements Expression {
     private final Pattern expressionRegexp;
     private final ParameterTypeRegistry parameterTypeRegistry;
     private final TreeRegexp treeRegexp;